+++ app/src/main/java/com/wear/pin/data/remote/pinterest/mapper/AuthMapper.kt
@@ -1,0 +1,5 @@
+package com.wear.pin.data.remote.pinterest.mapper
+
+object AuthMapper {
+    // TODO: Implement DTO to Domain mapping for OAuthToken
+}