+++ app/src/main/java/com/wear/pin/data/remote/pinterest/mapper/BoardMapper.kt
@@ -1,0 +1,5 @@
+package com.wear.pin.data.remote.pinterest.mapper
+
+object BoardMapper {
+    // TODO: Implement DTO to Domain mapping for Board
+}