+++ app/src/main/java/com/wear/pin/data/remote/pinterest/mapper/PinMapper.kt
@@ -1,0 +1,5 @@
+package com.wear.pin.data.remote.pinterest.mapper
+
+object PinMapper {
+    // TODO: Implement DTO to Domain mapping for Pin
+}