+++ app/src/main/java/com/wear/pin/data/remote/pinterest/mapper/UserMapper.kt
@@ -1,0 +1,5 @@
+package com.wear.pin.data.remote.pinterest.mapper
+
+object UserMapper {
+    // TODO: Implement DTO to Domain mapping for UserAccount
+}