+++ app/src/main/java/com/wear/pin/domain/repository/AuthRepository.kt
@@ -1,0 +1,5 @@
+package com.wear.pin.domain.repository
+
+interface AuthRepository {
+    // TODO: Define authentication methods
+}