+++ app/src/main/java/com/wear/pin/domain/repository/BoardRepository.kt
@@ -1,0 +1,5 @@
+package com.wear.pin.domain.repository
+
+interface BoardRepository {
+    // TODO: Define board-related methods
+}