+++ app/src/main/java/com/wear/pin/domain/repository/PinRepository.kt
@@ -1,0 +1,5 @@
+package com.wear.pin.domain.repository
+
+interface PinRepository {
+    // TODO: Define pin-related methods
+}