+++ app/src/main/java/com/wear/pin/domain/repository/UserRepository.kt
@@ -1,0 +1,5 @@
+package com.wear.pin.domain.repository
+
+interface UserRepository {
+    // TODO: Define user account methods
+}